--- app/javascript/components/widgets/forest-change/tree-gain-located/settings.js
@@ -1,8 +1,0 @@
-export default {
-  threshold: 50,
-  unit: 'ha',
-  pageSize: 5,
-  page: 0,
-  extentYear: 2000,
-  ifl: 2000
-};
