--- app/javascript/components/widgets/forest-change/tree-loss-located/settings.js
@@ -1,10 +1,0 @@
-export default {
-  threshold: 30,
-  extentYear: 2000,
-  unit: '%',
-  pageSize: 5,
-  page: 0,
-  startYear: 2001,
-  endYear: 2018,
-  ifl: 2000
-};
