--- app/javascript/components/widgets/forest-change/tree-loss-plantations/settings.js
@@ -1,6 +1,0 @@
-export default {
-  threshold: 30,
-  startYear: 2013,
-  endYear: 2018,
-  extentYear: 2010
-};
